feat: move excel templating to global service file
