--- backend/Controllers/Quotation/QuotationModule.cs
@@ -1,10 +1,0 @@
-namespace PeruControl.Controllers;
-
-public class QuotationModule : IModule
-{
-    public IServiceCollection SetupModule(IServiceCollection services, IConfiguration configuration)
-    {
-        services.AddScoped<QuotationService>();
-        return services;
-    }
-}
